chore: expose filename when panic
